requirements.txt
pandas
sqlalchemy
pymysql
python-dotenv
pyyaml
openpyxl
requests
boto3
urllib3